Replace TAB by space.
This commit was SVN r13269.

+++ opal/util/output.c
@@ -675,14 +675,14 @@ static void output(int output_id, const char *format, va_list arglist)
 	/* stdout output */
 
 	if (ldi->ldi_stdout) {
-            write(fileno(stdout), temp_str, strlen(temp_str)); 
+            write(fileno(stdout), temp_str, (int)strlen(temp_str)); 
             fflush(stdout);
 	}
 
 	/* stderr output */
 
 	if (ldi->ldi_stderr) {
-            write(fileno(stderr),temp_str, strlen(temp_str)); 
+            write(fileno(stderr),temp_str, (int)strlen(temp_str)); 
             fflush(stderr);
 	}
 
@@ -698,15 +698,15 @@ static void output(int output_id, const char *format, va_list arglist)
 		    } else if (ldi->ldi_file_num_lines_lost > 0) {
 		        char buffer[BUFSIZ];
 		        memset(buffer, 0, BUFSIZ);
-		        snprintf(buffer, BUFSIZ - 1,
-			             "[WARNING: %d lines lost because the Open MPI process session directory did\n not exist when opal_output() was invoked]\n",
-			             ldi->ldi_file_num_lines_lost);
-		        write(ldi->ldi_fd, buffer, strlen(buffer));
+                        snprintf(buffer, BUFSIZ - 1,
+                                 "[WARNING: %d lines lost because the Open MPI process session directory did\n not exist when opal_output() was invoked]\n",
+                                 ldi->ldi_file_num_lines_lost);
+		        write(ldi->ldi_fd, buffer, (int)strlen(buffer));
 		        ldi->ldi_file_num_lines_lost = 0;
 		    }
 	    }
 	    if (ldi->ldi_fd != -1) {
-            write(ldi->ldi_fd, temp_str, total_len);
+                write(ldi->ldi_fd, temp_str, (int)total_len);
 	    }
 	}
 	OPAL_THREAD_UNLOCK(&mutex);
